requirements.txt
PySide6
python-dotenv
openai-whisper
torch
opencv-python
pillow
requests
# For OCR
pytesseract
# For LLMs (OpenAI API, etc.)
openai
# For environment management
python-dotenv
# For audio
sounddevice
# For speech recognition
SpeechRecognition
# For PDF/image OCR (optional)
pdf2image
# For web search (optional)
serpapi
# Add any other dependencies as needed 